+++ src/vector/arithmetic/vector_mul_dd.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void dd_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -41,7 +41,7 @@ void dd_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::mul(dd_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -79,7 +79,7 @@ void dd_real_vector::mul(d_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::mul(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
