+++ src/vector/arithmetic/vector_sub_d.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void d_real_vector::sub(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -41,7 +41,7 @@ void d_real_vector::sub(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void d_real_vector::sub(d_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -75,7 +75,7 @@ void d_real_vector::sub(d_real_vector& vec1, dd_real_vector& vec2)
 
 void d_real_vector::sub(dd_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -110,13 +110,13 @@ void d_real_vector::sub(dd_real_vector& vec1, d_real_vector& vec2)
 
 void d_real_vector::sub(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cout << "error vecvor size is" << size() << vec1.size() << vec2.size() << std::endl;
 		assert(1);
 	}
 
 #pragma omp parallel for
-	for(size_t i = 0; i < (size_t)size(); i++){
+	for(size_t i = 0; i < size(); i++){
 		data()[i] = vec1.data()[i] - vec2.data()[i];
 	}
 }
