+++ src/vector/blas/dot.cpp
@@ -4,7 +4,7 @@ using namespace ddavx_core;
 namespace dd_avx{
 
 	dd_real dot(const dd_real_vector& x, const dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -16,7 +16,7 @@ namespace dd_avx{
 		{
 			int thN = omp_get_thread_num();
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 
 			AVXreg r_hi = regs.zeros;
 			AVXreg r_lo = regs.zeros;
@@ -50,7 +50,7 @@ namespace dd_avx{
 
 
 	dd_real dot(const d_real_vector& x, const dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -62,7 +62,7 @@ namespace dd_avx{
 		{
 			int thN = omp_get_thread_num();
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 
 			AVXreg r_hi = regs.zeros;
 			AVXreg r_lo = regs.zeros;
@@ -94,7 +94,7 @@ namespace dd_avx{
 	}
 
 	dd_real dot(const dd_real_vector& x, const d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -106,7 +106,7 @@ namespace dd_avx{
 		{
 			int thN = omp_get_thread_num();
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 
 			AVXreg r_hi = regs.zeros;
 			AVXreg r_lo = regs.zeros;
@@ -138,7 +138,7 @@ namespace dd_avx{
 	}
 
 	dd_real dot(const d_real_vector& x, const d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -150,7 +150,7 @@ namespace dd_avx{
 		{
 			int thN = omp_get_thread_num();
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 
 			AVXreg r_hi = regs.zeros;
 			AVXreg r_lo = regs.zeros;
