+++ src/vector/blas/scale.cpp
@@ -8,7 +8,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)x.size(), is, ie);
+			get_isie(x.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -33,7 +33,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)x.size(), is, ie);
+			get_isie(x.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -58,7 +58,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)x.size(), is, ie);
+			get_isie(x.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -82,7 +82,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)x.size(), is, ie);
+			get_isie(x.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
